[core] Moved the calculateMaxCasts function that was originally in the now defunct getCastEfficiency file into its own file
