+++ src/Parser/Warlock/Demonology/Modules/Features/Felstorm.js
@@ -3,7 +3,7 @@ import React from 'react';
 import Analyzer from 'Parser/Core/Analyzer';
 
 import SPELLS from 'common/SPELLS';
-import { calculateMaxCasts } from 'Parser/Core/getCastEfficiency';
+import calculateMaxCasts from 'Parser/Core/calculateMaxCasts';
 import SpellLink from 'common/SpellLink';
 import { formatPercentage } from 'common/format';
