+++ src/Parser/Warlock/Destruction/Modules/Features/DoomguardInfernal.js
@@ -4,7 +4,7 @@ import Analyzer from 'Parser/Core/Analyzer';
 import AbilityTracker from 'Parser/Core/Modules/AbilityTracker';
 import Combatants from 'Parser/Core/Modules/Combatants';
 
-import { calculateMaxCasts } from 'Parser/Core/getCastEfficiency';
+import calculateMaxCasts from 'Parser/Core/calculateMaxCasts';
 import SPELLS from 'common/SPELLS';
 import SpellLink from 'common/SpellLink';
